Revert "created new endpoint to wrap azure blob endpoint (#696)"
This reverts commit 569206e91b66edbbd4aa6c2ba0c5fdb2725cb137.
